updated cortex analyst
