+++ agents/cortex-analyst-example/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.0] - 2025-03-12
+
+- Update to use the Cortex Analyst action package instead of generic Snowflake package
+
 ## [0.0.6] - 2025-03-12
 
 - Add exclude file rules to `agent-spec.yaml`
