+++ agents/cortex-analyst-example/actions/Sema4.ai/snowflake-cortex-analyst/1.0.0/.gitignore
@@ -1,0 +1,13 @@
+*.pyc
+*.zip
+.DS_Store
+.env
+.project
+.pydevproject
+.use
+.venv/
+.vscode
+metadata.json
+output/
+temp/
+venv/
