+++ agents/cortex-analyst-example/actions/Sema4.ai/snowflake-cortex-analyst/1.0.0/utils.py
