--- agents/cortex-analyst-example/actions/Sema4.ai/snowflake-cortex/devdata/input_snowflake_execute_query.json
@@ -1,29 +1,0 @@
-{
-    "inputs": [
-        {
-            "inputName": "input-1",
-            "inputValue": {
-                "query": "SELECT * FROM production_reports LIMIT 10",
-                "warehouse": "COMPUTE_WH",
-                "database": "PRODUCTION_RESULTS",
-                "schema": "PUBLIC",
-                "numeric_args": null
-            }
-        }
-    ],
-    "metadata": {
-        "actionName": "snowflake_execute_query",
-        "actionRelativePath": "snowflake_actions.py",
-        "schemaDescription": [
-            "query: string: The query to execute.",
-            "warehouse: string: The warehouse to use.",
-            "database: string: The database to use.",
-            "schema: string: The schema to use.",
-            "numeric_args: string: A list of numeric arguments to pass to the query."
-        ],
-        "managedParamsSchemaDescription": {},
-        "inputFileVersion": "v3",
-        "kind": "action",
-        "actionSignature": "action/args: \"query: str, warehouse: str='', database: str=None, schema: str=None, numeric_args: list=None\""
-    }
-}
